+++ cli/actions/slashing.go
@@ -41,22 +41,25 @@ type SlashingImportOptions struct {
 	StartValidator  bool
 	GenerationPath  string
 	From            string
+	ContainerTag    string
 }
 
 func (s *sedgeActions) ImportSlashingInterchangeData(options SlashingImportOptions) error {
+	validatorContainerName := services.ContainerNameWithTag(services.DefaultSedgeValidatorClient, options.ContainerTag)
+	slashingContainerName := services.ContainerNameWithTag(services.ServiceCtSlashingData, options.ContainerTag)
 	// Check validator container exists
-	_, err := s.serviceManager.ContainerId(services.ServiceCtValidator)
+	_, err := s.serviceManager.ContainerId(validatorContainerName)
 	if err != nil {
 		return err
 	}
-	previouslyRunning, err := s.serviceManager.IsRunning(services.ServiceCtValidator)
+	previouslyRunning, err := s.serviceManager.IsRunning(validatorContainerName)
 	if err != nil {
 		return err
 	}
 	// Stop validator
 	if previouslyRunning {
 		log.Info("Stopping validator client...")
-		if err := s.serviceManager.Stop(services.ServiceCtValidator); err != nil {
+		if err := s.serviceManager.Stop(validatorContainerName); err != nil {
 			return err
 		}
 		log.Info("Validator client stopped.")
@@ -105,14 +108,14 @@ func (s *sedgeActions) ImportSlashingInterchangeData(options SlashingImportOptio
 		return fmt.Errorf("%w: %s", UnsupportedValidatorClientError, options.ValidatorClient)
 	}
 	log.Infof("Importing slashing data to client %s from %s", options.ValidatorClient, options.From)
-	if err := runSlashingContainer(s.dockerClient, s.serviceManager, cmd); err != nil {
+	if err := runSlashingContainer(s.dockerClient, s.serviceManager, cmd, validatorContainerName, slashingContainerName); err != nil {
 		return err
 	}
 
 	// Run validator again
 	if (previouslyRunning && !options.StopValidator) || options.StartValidator {
 		log.Info("The validator container is being restarted")
-		if err := s.serviceManager.Start(services.ServiceCtValidator); err != nil {
+		if err := s.serviceManager.Start(validatorContainerName); err != nil {
 			return err
 		}
 		log.Info("Validator started.")
@@ -127,22 +130,25 @@ type SlashingExportOptions struct {
 	StartValidator  bool
 	GenerationPath  string
 	Out             string
+	ContainerTag    string
 }
 
 func (s *sedgeActions) ExportSlashingInterchangeData(options SlashingExportOptions) error {
+	validatorContainerName := services.ContainerNameWithTag(services.DefaultSedgeValidatorClient, options.ContainerTag)
+	slashingContainerName := services.ContainerNameWithTag(services.ServiceCtSlashingData, options.ContainerTag)
 	// Check validator container exists
-	_, err := s.serviceManager.ContainerId(services.ServiceCtValidator)
+	_, err := s.serviceManager.ContainerId(validatorContainerName)
 	if err != nil {
 		return err
 	}
-	previouslyRunning, err := s.serviceManager.IsRunning(services.ServiceCtValidator)
+	previouslyRunning, err := s.serviceManager.IsRunning(validatorContainerName)
 	if err != nil {
 		return err
 	}
 	// Stop validator client
 	if previouslyRunning {
 		log.Info("Stopping validator client")
-		if err := s.serviceManager.Stop(services.ServiceCtValidator); err != nil {
+		if err := s.serviceManager.Stop(validatorContainerName); err != nil {
 			return err
 		}
 		log.Info("Validator client stopped.")
@@ -184,7 +190,7 @@ func (s *sedgeActions) ExportSlashingInterchangeData(options SlashingExportOptio
 		return fmt.Errorf("%w: %s", UnsupportedValidatorClientError, options.ValidatorClient)
 	}
 	log.Infof("Exporting slashing data from client %s", options.ValidatorClient)
-	if err := runSlashingContainer(s.dockerClient, s.serviceManager, cmd); err != nil {
+	if err := runSlashingContainer(s.dockerClient, s.serviceManager, cmd, validatorContainerName, slashingContainerName); err != nil {
 		return err
 	}
 	copyFrom := filepath.Join(options.GenerationPath, configs.ValidatorDir, "slashing_protection.json")
@@ -196,7 +202,7 @@ func (s *sedgeActions) ExportSlashingInterchangeData(options SlashingExportOptio
 	// Run validator again
 	if (previouslyRunning && !options.StopValidator) || options.StartValidator {
 		log.Info("The validator container is being restarted...")
-		if err := s.serviceManager.Start(services.ServiceCtValidator); err != nil {
+		if err := s.serviceManager.Start(validatorContainerName); err != nil {
 			return err
 		}
 		log.Info("Validator started.")
@@ -204,8 +210,10 @@ func (s *sedgeActions) ExportSlashingInterchangeData(options SlashingExportOptio
 	return nil
 }
 
-func runSlashingContainer(dockerClient client.APIClient, serviceManager services.ServiceManager, cmd []string) error {
-	validatorImage, err := serviceManager.Image(services.ServiceCtValidator)
+func runSlashingContainer(dockerClient client.APIClient, serviceManager services.ServiceManager, cmd []string,
+	validatorContainerName string, slashingContainerName string,
+) error {
+	validatorImage, err := serviceManager.Image(validatorContainerName)
 	if err != nil {
 		return err
 	}
@@ -216,17 +224,17 @@ func runSlashingContainer(dockerClient client.APIClient, serviceManager services
 			Cmd:   cmd,
 		},
 		&container.HostConfig{
-			VolumesFrom: []string{services.ServiceCtValidator},
+			VolumesFrom: []string{validatorContainerName},
 		},
 		&network.NetworkingConfig{},
 		&v1.Platform{},
-		services.ServiceCtSlashingData,
+		slashingContainerName,
 	)
 	if err != nil {
 		return err
 	}
 	log.Debugf("Slashing protection container id: %s", ct.ID)
-	ctExit, errChan := serviceManager.Wait(services.ServiceCtSlashingData, container.WaitConditionNextExit)
+	ctExit, errChan := serviceManager.Wait(slashingContainerName, container.WaitConditionNextExit)
 	log.Info("The slashing protection container is starting...")
 	if err := dockerClient.ContainerStart(context.Background(), ct.ID, types.ContainerStartOptions{}); err != nil {
 		return err