+++ cli/actions/slashing_test.go
@@ -50,6 +50,22 @@ func TestSlashingExport_ValidatorNotFound(t *testing.T) {
 	assert.ErrorIs(t, err, services.ErrContainerNotFound)
 }
 
+func validatorNotFoundHelper(t *testing.T, ctrl *gomock.Controller) actions.SedgeActions {
+	t.Helper()
+	dockerClient := mock_client.NewMockAPIClient(ctrl)
+
+	dockerClient.EXPECT().
+		ContainerList(gomock.Any(), types.ContainerListOptions{
+			All:     true,
+			Filters: filters.NewArgs(filters.Arg("name", services.DefaultSedgeValidatorClient)),
+		}).
+		Return(make([]types.Container, 0), nil).
+		Times(1)
+
+	serviceManager := services.NewServiceManager(dockerClient)
+	return actions.NewSedgeActions(dockerClient, serviceManager, nil)
+}
+
 func TestSlashingImport_CheckValidatorFailure(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	defer ctrl.Finish()
@@ -72,6 +88,27 @@ func TestSlashingExport_CheckValidatorFailure(t *testing.T) {
 	assert.ErrorIs(t, err, wantError)
 }
 
+func checkValidatorFailureHelper(t *testing.T, ctrl *gomock.Controller, wantError error) actions.SedgeActions {
+	dockerClient := mock_client.NewMockAPIClient(ctrl)
+
+	dockerClient.EXPECT().
+		ContainerList(gomock.Any(), types.ContainerListOptions{
+			All:     true,
+			Filters: filters.NewArgs(filters.Arg("name", services.DefaultSedgeValidatorClient)),
+		}).
+		Return([]types.Container{
+			{ID: "validatorctid"},
+		}, nil).
+		Times(1)
+	dockerClient.EXPECT().
+		ContainerInspect(gomock.Any(), services.DefaultSedgeValidatorClient).
+		Return(types.ContainerJSON{}, wantError).
+		Times(1)
+
+	serviceManager := services.NewServiceManager(dockerClient)
+	return actions.NewSedgeActions(dockerClient, serviceManager, nil)
+}
+
 func TestSlashingImport_ValidatorStopFailure(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	defer ctrl.Finish()
@@ -92,6 +129,40 @@ func TestSlashingExport_ValidatorStopFailure(t *testing.T) {
 	assert.ErrorIs(t, err, services.ErrStoppingContainer)
 }
 
+func validatorStopFailureHelper(t *testing.T, ctrl *gomock.Controller) actions.SedgeActions {
+	dockerClient := mock_client.NewMockAPIClient(ctrl)
+
+	validatorCtId := "validatorctid"
+
+	dockerClient.EXPECT().
+		ContainerList(gomock.Any(), types.ContainerListOptions{
+			All:     true,
+			Filters: filters.NewArgs(filters.Arg("name", services.DefaultSedgeValidatorClient)),
+		}).
+		Return([]types.Container{
+			{ID: validatorCtId},
+		}, nil).
+		Times(1)
+	dockerClient.EXPECT().
+		ContainerInspect(gomock.Any(), services.DefaultSedgeValidatorClient).
+		Return(types.ContainerJSON{
+			ContainerJSONBase: &types.ContainerJSONBase{
+				ID: validatorCtId,
+				State: &types.ContainerState{
+					Running: true,
+				},
+			},
+		}, nil).
+		Times(2)
+	dockerClient.EXPECT().
+		ContainerStop(gomock.Any(), "validatorctid", gomock.Any()).
+		Return(errors.New("error")).
+		Times(1)
+
+	serviceManager := services.NewServiceManager(dockerClient)
+	return actions.NewSedgeActions(dockerClient, serviceManager, nil)
+}
+
 func TestSlashingImport_ValidatorRunning(t *testing.T) {
 	clients := []string{"prysm", "lighthouse", "lodestar", "teku"}
 	for _, validatorClient := range clients {
@@ -99,7 +170,7 @@ func TestSlashingImport_ValidatorRunning(t *testing.T) {
 			ctrl := gomock.NewController(t)
 			defer ctrl.Finish()
 
-			s := slashingGoldenPath(t, ctrl)
+			s := slashingGoldenPath(t, ctrl, "")
 
 			generationPath := t.TempDir()
 			from := setupSlashingDataFile(t, filepath.Join(t.TempDir(), "slashing-data.json"))
@@ -128,7 +199,9 @@ func TestSlashingExport_ValidatorRunning(t *testing.T) {
 			ctrl := gomock.NewController(t)
 			defer ctrl.Finish()
 
-			s := slashingGoldenPath(t, ctrl)
+			tag := "sampleTag"
+
+			s := slashingGoldenPath(t, ctrl, tag)
 
 			generationPath := t.TempDir()
 
@@ -139,6 +212,7 @@ func TestSlashingExport_ValidatorRunning(t *testing.T) {
 				Network:         "sepolia",
 				GenerationPath:  generationPath,
 				Out:             out,
+				ContainerTag:    tag,
 			})
 			assert.Nil(t, err)
 			assert.FileExists(t, out)
@@ -154,18 +228,19 @@ func TestSlashingExport_ValidatorRunning(t *testing.T) {
 // slashingGoldenPath returns a SedgeActions interface with a mocked docker client
 // with all the required responses for a correct slashing container execution.
 // This setup is valid for the export and import process.
-func slashingGoldenPath(t *testing.T, ctrl *gomock.Controller) actions.SedgeActions {
+func slashingGoldenPath(t *testing.T, ctrl *gomock.Controller, containerTag string) actions.SedgeActions {
 	dockerClient := mock_client.NewMockAPIClient(ctrl)
 
 	validatorCtId := "validatorctid"
-	slashingCtName := "validator-slashing-data"
+	slashingCtName := services.ContainerNameWithTag(services.ServiceCtSlashingData, containerTag)
 	slashingCtId := "slashing-ct-id"
 
+	validatorContainerName := services.ContainerNameWithTag(services.DefaultSedgeValidatorClient, containerTag)
 	// Mock ContainerList
 	dockerClient.EXPECT().
 		ContainerList(gomock.Any(), types.ContainerListOptions{
 			All:     true,
-			Filters: filters.NewArgs(filters.Arg("name", services.ServiceCtValidator)),
+			Filters: filters.NewArgs(filters.Arg("name", validatorContainerName)),
 		}).
 		Return([]types.Container{
 			{ID: validatorCtId},
@@ -173,7 +248,7 @@ func slashingGoldenPath(t *testing.T, ctrl *gomock.Controller) actions.SedgeActi
 		Times(1)
 	// Mock ContainerInspect
 	dockerClient.EXPECT().
-		ContainerInspect(gomock.Any(), services.ServiceCtValidator).
+		ContainerInspect(gomock.Any(), validatorContainerName).
 		Return(types.ContainerJSON{
 			ContainerJSONBase: &types.ContainerJSONBase{
 				ID: validatorCtId,
@@ -199,7 +274,7 @@ func slashingGoldenPath(t *testing.T, ctrl *gomock.Controller) actions.SedgeActi
 		Return(nil).
 		Times(1)
 	dockerClient.EXPECT().
-		ContainerStart(gomock.Any(), services.ServiceCtValidator, gomock.Any()).
+		ContainerStart(gomock.Any(), validatorContainerName, gomock.Any()).
 		Return(nil).
 		Times(1)
 	// Mock ContainerWait
@@ -284,7 +359,7 @@ func unsupportedClientsHelper(t *testing.T, ctrl *gomock.Controller) actions.Sed
 	dockerClient.EXPECT().
 		ContainerList(gomock.Any(), types.ContainerListOptions{
 			All:     true,
-			Filters: filters.NewArgs(filters.Arg("name", services.ServiceCtValidator)),
+			Filters: filters.NewArgs(filters.Arg("name", services.DefaultSedgeValidatorClient)),
 		}).
 		Return([]types.Container{
 			{ID: validatorCtId},
@@ -292,7 +367,7 @@ func unsupportedClientsHelper(t *testing.T, ctrl *gomock.Controller) actions.Sed
 		Times(1)
 	// Mock ContainerInspect
 	dockerClient.EXPECT().
-		ContainerInspect(gomock.Any(), services.ServiceCtValidator).
+		ContainerInspect(gomock.Any(), services.DefaultSedgeValidatorClient).
 		Return(types.ContainerJSON{
 			ContainerJSONBase: &types.ContainerJSONBase{
 				ID: validatorCtId,